MS-07: Organize components and modules.css in folders as requested
