+++ src/Components/Sessions/ConfirmModal/confirmModal.module.css
@@ -1,0 +1,48 @@
+.modal {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  position: absolute;
+  width: 100%;
+  height: 500px;
+  background-color: rgba(0,0,0,0.4);
+  z-index: 1;
+}
+.confirmModalContent {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 60%;
+  background-color: white;
+}
+.btnContainer {
+  display: flex;
+  justify-content: space-between;
+  width: 40%;
+  padding-bottom: 20px;
+}
+.btnContainer button {
+  padding: 10px 50px;
+}
+.confirmBtn, .cancelBtn {
+  background-color: #fc692f;
+  border: none;
+  border-radius: 3px;
+  color: white;
+  cursor: pointer;
+}
+.cancelBtn {
+  background-color: #232f3e;
+}
+.modalHeader {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 60%;
+  background-color: #232f3e;
+  color: white;
+}
+.modalHeader * {
+  padding: 0 40px;
+}
