+++ src/Components/Sessions/CreateModal/index.js
@@ -1,4 +1,4 @@
-import styles from './sessions.module.css';
+import styles from './createModal.module.css';
 import { useEffect, useState } from 'react';
 import { FaTimes } from 'react-icons/fa';
