--- src/Components/Sessions/Header.js
@@ -1,5 +1,0 @@
-const Header = () => {
-  return <h2>Sessions</h2>;
-};
-
-export default Header;
