--- src/Components/Sessions/ListHeader.js
@@ -1,16 +1,0 @@
-const ListHeader = () => {
-  return (
-    <thead>
-      <tr>
-        <th>Postulant</th>
-        <th>Psychologist</th>
-        <th>Status</th>
-        <th>Date</th>
-        <th>Actions</th>
-        <th></th>
-      </tr>
-    </thead>
-  );
-};
-
-export default ListHeader;
