+++ src/Components/Sessions/Session/session.module.css
@@ -1,0 +1,16 @@
+table {
+  width: 100%;
+  text-align: left;
+  border-collapse: collapse;
+}
+tr {
+  border-bottom: 1px solid grey;
+  font-weight: 600;
+}
+th {
+  color: #fc692f;
+}
+svg {
+  margin: 5px;
+  cursor: pointer;
+}
