+++ src/Components/Sessions/UpdateModal/updateModal.module.css
@@ -1,0 +1,57 @@
+.modal {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  position: absolute;
+  width: 100%;
+  height: 500px;
+  background-color: rgba(0,0,0,0.4);
+  z-index: 1;
+}
+.createForm {
+  background-color: white;
+}
+form {
+  display: flex;
+  justify-content: space-around;
+  flex-wrap: wrap;
+  width: 60%;
+  padding-top: 20px;
+}
+.inputContainer {
+  display: flex;
+  align-items: center;
+  width: 50%;
+}
+label {
+  display: flex;
+  flex-direction: column;
+  width: fit-content;
+  margin-left: 15%;
+}
+input, select {
+  background: white;
+  width: 200px;
+  height: 40px;
+  border: 1px solid gray;
+  border-radius: 3px;
+  margin-bottom: 20px;
+}
+.submitBtn {
+  border: none;
+  color: white;
+  background-color: #fc692f;
+  cursor: pointer;
+}
+.modalHeader {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 60%;
+  background-color: #232f3e;
+  color: white;
+}
+.modalHeader * {
+  padding: 0 40px;
+}
