+++ src/Components/Sessions/sessions.module.css
@@ -1,20 +1,6 @@
 .container {
   height: 500px
 }
-.modal {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  position: absolute;
-  width: 100%;
-  height: 500px;
-  background-color: rgba(0,0,0,0.4);
-  z-index: 1;
-}
-.createForm {
-  background-color: white;
-}
 table {
   width: 100%;
   text-align: left;
@@ -40,75 +26,6 @@ svg {
   background-color: transparent;
   color: #fc692f;
   cursor: pointer;
-}
-.confirmModalContent {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 60%;
-  background-color: white;
-}
-.btnContainer {
-  display: flex;
-  justify-content: space-between;
-  width: 40%;
-  padding-bottom: 20px;
-}
-.btnContainer button {
-  padding: 10px 50px;
-}
-.confirmBtn, .cancelBtn {
-  background-color: #fc692f;
-  border: none;
-  border-radius: 3px;
-  color: white;
-  cursor: pointer;
-}
-.cancelBtn {
-  background-color: #232f3e;
-}
-form {
-  display: flex;
-  justify-content: space-around;
-  flex-wrap: wrap;
-  width: 60%;
-  padding-top: 20px;
-}
-.inputContainer {
-  display: flex;
-  align-items: center;
-  width: 50%;
-}
-label {
-  display: flex;
-  flex-direction: column;
-  width: fit-content;
-  margin-left: 15%;
-}
-input, select {
-  background: white;
-  width: 200px;
-  height: 40px;
-  border: 1px solid gray;
-  border-radius: 3px;
-  margin-bottom: 20px;
-}
-.submitBtn {
-  border: none;
-  color: white;
-  background-color: #fc692f;
-  cursor: pointer;
-}
-.modalHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 60%;
-  background-color: #232f3e;
-  color: white;
-}
-.modalHeader * {
-  padding: 0 40px;
 }
 section {
   width: 90%;
